Sort method fields in resources object
The code has been updated to not only delete resources with a count of zero, but also to sort the fields in methods object of each resource in ascending order. The sorting is based on the summary attribute of the operation object. This will enhance readability and provide a systematic view of the methods within each resource.